+++ src/stac_asset/planetary_computer_client.py
@@ -50,7 +50,7 @@ class _Token:
 
 
 class PlanetaryComputerClient(HttpClient):
-    cache: Dict[URL, _Token]
+    _cache: Dict[URL, _Token]
     token_request_url: URL
 
     def __init__(
@@ -59,7 +59,7 @@ class PlanetaryComputerClient(HttpClient):
         sas_token_endpoint: str = DEFAULT_SAS_TOKEN_ENDPOINT,
     ) -> None:
         super().__init__(session)
-        self.cache = dict()
+        self._cache = dict()
         self.sas_token_endpoint = URL(sas_token_endpoint)
 
     async def open_url(self, url: URL) -> AsyncIterator[bytes]:
@@ -88,12 +88,12 @@ class PlanetaryComputerClient(HttpClient):
 
     async def _get_token(self, account_name: str, container_name: str) -> str:
         url = self.sas_token_endpoint.joinpath(account_name, container_name)
-        token = self.cache.get(url)
+        token = self._cache.get(url)
         if token is None or token.ttl() < 60:
             response = await self.session.get(url)
             response.raise_for_status()
             token = _Token.from_dict(await response.json())
-            self.cache[url] = token
+            self._cache[url] = token
         return str(token)
 
     async def __aenter__(self) -> PlanetaryComputerClient: